feat: frontend interfaces
