+++ src/frontend/__init__.py
@@ -1,9 +1,11 @@
+from .interfaces import ModalBase, PresenceChangeable
 from .components import Chat, ConnBox
 from .agenda_frame import AgendaFrame
 from .chat_frame import ChatFrame
 from .layout import Layout
 from .modal import Modal
 
-__all__ = ["Chat", "ConnBox",
+__all__ = ["ModalBase", "PresenceChangeable",
+           "Chat", "ConnBox",
            "AgendaFrame", "ChatFrame", "Layout",
            "Modal"]
