+++ src/frontend/components/__init__.py
@@ -1,5 +1,4 @@
 from .chat import Chat
 from .conn_box import ConnBox
-from .interfaces import PresenceChangeable
 
-__all__ = ["Chat", "ConnBox", "PresenceChangeable"]
+__all__ = ["Chat", "ConnBox"]
