+++ src/frontend/interfaces.py
@@ -17,3 +17,22 @@ class ModalBase(ft.Container):
             width=450,
             height=600
         )
+
+class PresenceChangeable(ft.Control):
+    """
+    Controls that live within a page and can show/hide themselves.
+    """
+    
+    def show(self):
+        """
+        Makes the control visible and updates the page.
+        """
+        self.visible = True
+        self.update()
+    
+    def hide(self):
+        """
+        Makes the control invisible and updates the page.
+        """
+        self.visible = False
+        self.update()
